pyproject.toml
[project]
name = "gmail-pubsub"
version = "0.1.0"
description = "Gmail Push Processing API with Cloud Run and Pub/Sub"
authors = [
    {name = "Developer", email = "developer@example.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Web framework
    "fastapi>=0.119.0",
    "uvicorn>=0.37.0",
    
    # Google Cloud libraries
    "google-cloud-pubsub>=2.18.0",
    "google-api-python-client>=2.100.0",
    "google-auth>=2.23.0",
    "google-auth-oauthlib>=1.1.0",
    "google-auth-httplib2>=0.2.0",
    "google-cloud-secret-manager>=2.16.0",
    
    # Database
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",  # PostgreSQL driver
    "aiosqlite>=0.19.0",       # Async SQLite support
    
    # Data validation and parsing
    "pydantic>=2.0.0",
    "beautifulsoup4>=4.14.2",
    
    # Configuration and utilities
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",  # For testing async functions
    "black>=23.7.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
    "httpx>=0.28.1",           # For testing HTTP endpoints
    "faker>=19.0.0",           # For generating test data
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app", "scripts"]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[dependency-groups]
dev = [
    "httpx>=0.28.1",
]

[project.scripts]
init = "scripts.init:main"
deploy = "scripts.deploy:main"
test = "scripts.test:main"
reset = "scripts.reset:main"
cleanup = "scripts.cleanup:main"
